+++ packages/ub/modelUB.js
@@ -312,6 +312,9 @@ UB.registerMixinModule('multitenancy', multitenancyImpl)
 const fsStorageImpl = require('./mixins/fsStorageMixin')
 UB.registerMixinModule('fsStorage', fsStorageImpl)
 
+const aclRlsStorageMixinImpl = require('./mixins/aclRlsStorageMixin')
+UB.registerMixinModule('aclRlsStorage', aclRlsStorageMixinImpl)
+
 if (App.serverConfig.security.disabledAccounts) {
   const reDisabledAccounts = new RegExp(App.serverConfig.security.disabledAccounts)
   Session.on('login', () => {